fix:[BE,FE] 화상채팅 css, 소캣 서버 통합
