+++ server/app.js
@@ -18,8 +18,6 @@ const server = http.createServer(app);
 const { swaggerUi, specs } = require("./modules/swagger/swagger");
 const socketHandler = require("./sockets");
 socketHandler(server);
-const socketWebRTC = require("./modules/webrtc/webrtc");
-socketWebRTC(server);
 
 const corsOptions = {
     origin: true,
